pyproject.toml
[project]
name = "stock-price-alerts-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.2",
    "fastapi[standard]>=0.133.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.6",
    "sqlalchemy>=2.0.47",
]
